requirements.txt
streamlit
requests
pandas
dnspython